Crate snake_bevy_wasm_pwa

snake_bevy_wasm_pwa

Template for a minimal pwa wasm project for browser
version: 0.0.12 date: 2025-08-13 author: bestia.dev repository: GitHub

this repo is part of the automation_tasks_rs project

Hashtags: #maintained #ready-for-use #rustlang #automation #workflow
My projects on GitHub are more like a tutorial than a finished product: bestia-dev tutorials.
I recommend using the CRUSTDE - Containerized Rust Development Environment to write Rust projects on Linux, isolated from your system.

This template

Just like cargo new makes a soft and gentle introduction to Rust projects and development, I want to make the same for an in-browser WASM project with

cargo auto new_pwa_wasm

Extremely simple, just the basic moving parts and use cases.
This simplest template does not have a PWA implementation or dedicated web server app.

Development details

Read the development details in a separate md file: DEVELOPMENT.md

Releases changelog

Read the releases changelog in a separate md file: RELEASES.md

TODO

And code happily ever afterโฆ

Open-source and free as a beer

My open-source projects are free as a beer (MIT license).
I just love programming.
But I need also to drink. If you find my projects and tutorials helpful, please buy me a beer by donating to my PayPal.
You know the price of a beer in your local bar ;-)
So I can drink a free beer for your health :-)
Na zdravje! Alla salute! Prost! Nazdravlje! ๐ป

//bestia.dev
//github.com/bestia-dev
//bestiadev.substack.com
//youtube.com/@bestia-dev-tutorials

Modules

main_mod ๐

This module is like a main.rs module for a binary CLI executable.
The main_mod.rs contains all input/output interface stuff.
So the program logic can be separate from the interface.

Enums

LibraryError

LibraryError must be accessible in every module. all possible library errors for thiserror

Functions

wasm_bindgen_start

To start the Wasm application, wasm_bindgen runs this functions