snake_bevy_wasm_pwa

Module main_mod

This module is like a main.rs module for a binary CLI executable.
The main_mod.rs contains all input/output interface stuff.
So the program logic can be separate from the interface.

Re-exports

pub use lib_mod::LibraryError;

Modules

lib_mod 🔒

This module is like a lib.rs module for a binary CLI executable. The lib_mod.rs must not contains any input/output interface stuff. So the program logic can be separate from the interface.

Functions

get_args_from_hash_fragment 🔒

get args from hash fragment

header 🔒

render header with Home and Help

main

entry point just like for cli-bin-executable

on_click_btn_run 🔒

the listener calls this function

on_hash_change 🔒

the listener calls this function

page_with_inputs 🔒

render first page

print_greet_name 🔒

print my name

print_help 🔒

print help

routing_by_arguments 🔒

routing by arguments
routing can come from:

upper_greet_name 🔒

print my name upper, can return error